requirements.txt
opencv-python
mtcnn
tqdm
torch
torchvision
pandas
pillow
scikit-learn
